[openstack] align recognized parameters, move common things among services into openstack core
