+++ lib/fog/openstack.rb
@@ -1,5 +1,6 @@
 require 'fog/openstack/compute'
 require 'fog/openstack/identity_v2'
+require 'fog/openstack/identity_v3'
 require 'fog/openstack/image'
 require 'fog/openstack/metering'
 require 'fog/openstack/network'
