+++ lib/fog/openstack/identity_v2.rb
@@ -7,11 +7,15 @@ module Fog
       class V2 < Fog::Service
 
         requires :openstack_auth_url
-        recognizes :openstack_auth_token, :openstack_management_url, :persistent,
-                   :openstack_service_type, :openstack_service_name, :openstack_tenant,
-                   :openstack_api_key, :openstack_username, :openstack_current_user_id,
-                   :current_user, :current_tenant,
-                   :openstack_endpoint_type, :openstack_region
+        recognizes :openstack_auth_token, :openstack_management_url,
+                   :persistent, :openstack_service_type, :openstack_service_name,
+                   :openstack_tenant, :openstack_tenant_id,
+                   :openstack_api_key, :openstack_username, :openstack_identity_endpoint,
+                   :current_user, :current_tenant, :openstack_region,
+                   :openstack_endpoint_type,
+                   :openstack_project_name, :openstack_project_id,
+                   :openstack_project_domain, :openstack_user_domain, :openstack_domain_name,
+                   :openstack_project_domain_id, :openstack_user_domain_id, :openstack_domain_id
 
         model_path 'fog/openstack/models/identity_v2'
         model :tenant
@@ -166,97 +170,23 @@ module Fog
         end
 
         class Real
-          attr_reader :current_user
-          attr_reader :current_tenant
-          attr_reader :unscoped_token
-
           include Fog::Identity::OpenStack::Common
 
           def initialize(options={})
-            @openstack_auth_token = options[:openstack_auth_token]
+            initialize_identity options
 
-            unless @openstack_auth_token
-              missing_credentials = Array.new
-              @openstack_api_key = options[:openstack_api_key]
-              @openstack_username = options[:openstack_username]
-              @openstack_region = options[:openstack_region]
+            @openstack_service_type   = options[:openstack_service_type] || ['identity']
+            @openstack_service_name   = options[:openstack_service_name]
 
-              missing_credentials << :openstack_api_key unless @openstack_api_key
-              missing_credentials << :openstack_username unless @openstack_username
-              raise ArgumentError, "Missing required arguments: #{missing_credentials.join(', ')}" unless missing_credentials.empty?
-            end
+            @connection_options       = options[:connection_options] || {}
 
-            @openstack_tenant = options[:openstack_tenant]
-            @openstack_auth_uri = URI.parse(options[:openstack_auth_url])
-            @openstack_management_url = options[:openstack_management_url]
-            @openstack_must_reauthenticate = false
-            @openstack_service_type = options[:openstack_service_type] || ['identity']
-            @openstack_service_name = options[:openstack_service_name]
-
-            @connection_options = options[:connection_options] || {}
-
-            @openstack_current_user_id = options[:openstack_current_user_id]
-
-            @openstack_endpoint_type = options[:openstack_endpoint_type] || 'adminURL'
-
-            @current_user = options[:current_user]
-            @current_tenant = options[:current_tenant]
+            @openstack_endpoint_type  = options[:openstack_endpoint_type] || 'adminURL'
 
             authenticate
 
             @persistent = options[:persistent] || false
             @connection = Fog::Core::Connection.new("#{@scheme}://#{@host}:#{@port}", @persistent, @connection_options)
           end
-
-          def credentials
-            {:provider => 'openstack',
-             :openstack_auth_url => @openstack_auth_uri.to_s,
-             :openstack_auth_token => @auth_token,
-             :openstack_management_url => @openstack_management_url,
-             :openstack_current_user_id => @openstack_current_user_id,
-             :current_user => @current_user,
-             :current_tenant => @current_tenant}
-          end
-
-          def reload
-            @connection.reset
-          end
-
-          def request(params)
-            retried = false
-            begin
-              response = @connection.request(params.merge({
-                                                              :headers => {
-                                                                  'Content-Type' => 'application/json',
-                                                                  'Accept' => 'application/json',
-                                                                  'X-Auth-Token' => @auth_token
-                                                              }.merge!(params[:headers] || {}),
-                                                              :path => "#{@path}/#{params[:path]}" #,
-                                                          }))
-            rescue Excon::Errors::Unauthorized => error
-              raise if retried
-              retried = true
-
-              @openstack_must_reauthenticate = true
-              authenticate
-              retry
-            rescue Excon::Errors::HTTPStatusError => error
-              raise case error
-                      when Excon::Errors::NotFound
-                        Fog::Identity::OpenStack::NotFound.slurp(error)
-                      else
-                        error
-                    end
-            end
-            unless response.body.empty?
-              response.body = Fog::JSON.decode(response.body)
-            end
-            response
-          end
-
-          private
-
-
         end
       end
     end
